application/src/main/java/org/example/service/EmailService.java
package org.example.service;

import lombok.AllArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.mail.SimpleMailMessage;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.stereotype.Service;

@Slf4j
@Service
@AllArgsConstructor
public class EmailService {

    private final JavaMailSender mailSender;

    public boolean sendEmail(
            String from,
            String to,
            String subject,
            String body
    ) {
        try {
            SimpleMailMessage message = new SimpleMailMessage();
            message.setFrom(from);
            message.setTo(to);
            message.setSubject(subject);
            message.setText(body);

            mailSender.send(message);
            return true; // Email sent successfully
        } catch (Exception e) {
            log.error("An error occurred while sending the email: {}", e.getMessage());
            return false; // Email sending failed
        }
    }
}
